pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aesthetic-indexer"
version = "0.1.0"
description = "Automated found-footage scraper + aesthetic indexer (clips, embeddings, tags, pgvector search)."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "delora" }]
dependencies = [
  "rich>=13.7.0",
  "tqdm>=4.66.0",
  "python-dotenv>=1.0.0",
  "typer>=0.12.0",
  "pydantic>=2.6.0",

  "yt-dlp>=2025.1.0",
  "internetarchive>=5.0.0",

  # PySceneDetect (package name is 'scenedetect')
  "scenedetect>=0.6.3",

  "opencv-python>=4.9.0.80",
  "numpy>=1.26.0",
  "scipy>=1.11.0",
  "pandas>=2.2.0",
  "Pillow>=10.2.0",
  "ImageHash>=4.3.1",

  "torch>=2.2.0",
  "open-clip-torch>=2.24.0",

  "librosa>=0.10.2",
  "soundfile>=0.12.1",
  "pydub>=0.25.1",

  "SQLAlchemy>=2.0.0",
  "psycopg2-binary>=2.9.9",
  "pgvector>=0.3.6",
  
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
]

[project.scripts]
ai-run-all = "src.pipeline.run_all:app"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]


